[ALPS03194731] Audio: Fix coverity issue
Fix unintentional integer overflow issue

MTK-Commit-Id: 936a3ba1b3ae354a0af067663587c5c47641b3f4

Feature: BesRecord
CR-Id: ALPS03194731

+++ sound/soc/mediatek/common_int/mtk-soc-pcm-btcvsd.c
@@ -716,7 +716,7 @@ ssize_t AudDrv_btcvsd_read(char __user *data, size_t count)
 
 	/* Save current timestamp & buffer time in bt_rx_timestamp and bt_rx_bufdata_equivalent_time */
 	bt_rx_timestamp = sched_clock();
-	bt_rx_bufdata_equivalent_time = (btsco.pRX->iPacket_w - btsco.pRX->iPacket_r) * (SCO_RX_PLC_SIZE)
+	bt_rx_bufdata_equivalent_time = (uint64)(btsco.pRX->iPacket_w - btsco.pRX->iPacket_r) * (SCO_RX_PLC_SIZE)
 					* 16 * 1000  / 2 / 64;
 	bt_rx_bufdata_equivalent_time += read_count * SCO_RX_PLC_SIZE * 16 * 1000
 					 / (SCO_RX_PLC_SIZE + BTSCO_CVSD_PACKET_VALID_SIZE) / 2 / 64;
@@ -748,7 +748,7 @@ ssize_t AudDrv_btcvsd_write(const char __user *data, size_t count)
 
 	/* Save current timestamp & buffer time in bt_tx_timestamp and bt_tx_bufdata_equivalent_time */
 	bt_tx_timestamp = sched_clock();
-	bt_tx_bufdata_equivalent_time = (btsco.pTX->iPacket_w - btsco.pTX->iPacket_r) * (SCO_TX_ENCODE_SIZE)
+	bt_tx_bufdata_equivalent_time = (uint64)(btsco.pTX->iPacket_w - btsco.pTX->iPacket_r) * (SCO_TX_ENCODE_SIZE)
 					* 16 * 1000  / 2 / 64;
 	bt_tx_bufdata_equivalent_time *= 1000; /* return equivalent time(us) to data count */
 	LOGBT("bt_tx_timestamp:%llu,bt_tx_bufdata_equivalent_time:%llu, iPacket_w:%d, iPacket_r:%d ",
